split-up fetcher; moved definitions to separate file
