+++ apps/researcher/src/lib/objects/searcher.integration.test.ts
@@ -1,4 +1,4 @@
-import {HeritageObjectsSearcher} from './objects-searcher';
+import {HeritageObjectSearcher} from './searcher';
 import {LabelFetcher} from '@colonial-collections/label-fetcher';
 import {beforeEach, describe, expect, it} from '@jest/globals';
 import {env} from 'node:process';
@@ -6,10 +6,10 @@ import {env} from 'node:process';
 const labelFetcher = new LabelFetcher({
   endpointUrl: env.SEARCH_PLATFORM_SPARQL_ENDPOINT_URL as string,
 });
-let heritageObjectsSearcher: HeritageObjectsSearcher;
+let heritageObjectSearcher: HeritageObjectSearcher;
 
 beforeEach(() => {
-  heritageObjectsSearcher = new HeritageObjectsSearcher({
+  heritageObjectSearcher = new HeritageObjectSearcher({
     endpointUrl: env.SEARCH_PLATFORM_ELASTIC_ENDPOINT_URL as string,
     labelFetcher,
   });
@@ -17,7 +17,7 @@ beforeEach(() => {
 
 describe('search', () => {
   it('finds all heritage objects if no options are provided', async () => {
-    const result = await heritageObjectsSearcher.search();
+    const result = await heritageObjectSearcher.search();
 
     expect(result).toStrictEqual({
       totalCount: 5,
